pyproject.toml
[project]
name = "entity-doctor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "pyhumps>=3.8.0",
    "rich>=14.0.0",
]
